Extract parser.dynamic
